The requirements for the file upload comments tasks were made

+++ app/Http/Controllers/TaskController.php
@@ -19,22 +19,21 @@ class TaskController extends Controller
      */
     public function index(): Response
     {
-        $users = User::all();
+        $tasks = Task::with('user:id,name', 'comments.user', 'attachments.user')->latest()->get();
 
         return Inertia::render('Tasks/Index', [
-            'tasks' => Task::with('user:id,name')->latest()->get(),
-            'users' => $users,
+            'tasks' => $tasks,
+            'users' => User::all(),
         ]);
     }
-
     /**
      * Show the form for creating a new resource.
      */
     public function create()
     {
         $this->authorize('create', Task::class);
-        
-        $users = User::all(); 
+
+        $users = User::all();
         return Inertia::render('Tasks/Create', [
             'users' => $users,
         ]);
@@ -64,6 +63,8 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        $task->load('comments.user');
+
         return view('tasks.show', compact('task'));
     }
 
@@ -75,34 +76,45 @@ class TaskController extends Controller
         return view('tasks.edit', compact('task'));
     }
 
+    public function updateStatus(Request $request, Task $task): RedirectResponse
+    {
+        $this->authorize('updateStatus', $task);
+
+        $validated = $request->validate([
+            'status' => 'required',
+        ]);
+        $task->update($validated);
+
+        return redirect(route('tasks.index'));
+    }
+    
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, Task $task): RedirectResponse
     {
-        // $request->validate([
-        //     'title' => 'required',
-        //     'description' => 'required',
-        //     'status' => 'required',
-        //     'user_id' => 'required',
-        // ]);
 
         Gate::authorize('update', $task);
 
         $validated = $request->validate([
             'title' => 'required',
+            'description' => 'required',
+            'status' => 'sometimes',
+            'user_id' => 'sometimes',
         ]);
-
         $task->update($validated);
 
         return redirect(route('tasks.index'));
     }
 
+
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Task $task)
+    public function destroy(Task $task): RedirectResponse
     {
+        Gate::authorize('delete', $task);
+
         $task->delete();
 
         return redirect()->route('tasks.index');
